Appropriately rename lib/subsystems folder to 'devices'
