+++ src/main/java/frc/robot/subsystems/leds/LEDs.java
@@ -17,8 +17,8 @@ package frc.robot.subsystems.leds;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.lib.devices.Lights;
 import frc.lib.io.lights.LightsIO;
-import frc.lib.subsystems.Lights;
 
 public class LEDs extends SubsystemBase {
     private final Lights lights;
